+++ spec/spec_helper.rb
@@ -1,18 +1,23 @@
 # frozen_string_literal: true
 
 require 'bundler/setup'
+require 'legion/logging'
+require 'legion/settings'
+require 'legion/json'
+require 'legion/cache'
+require 'legion/crypt'
+require 'legion/data'
+require 'legion/transport'
 require 'legion/extensions/exec'
 
-unless defined?(Legion::Logging)
+unless defined?(Legion::Extensions::Helpers::Lex)
   module Legion
-    module Logging
-      def self.info(*); end
-
-      def self.debug(*); end
-
-      def self.warn(*); end
-
-      def self.error(*); end
+    module Extensions
+      module Helpers
+        module Lex
+          def self.included(base); end
+        end
+      end
     end
   end
 end